+++ module/herosystem6e.js
@@ -48,6 +48,10 @@ Hooks.once('init', async function () {
     CONFIG.Combat.documentClass = HeroSystem6eCombat;
     CONFIG.Combat.defeatedStatusId = "dead";
 
+    // V11 now support ActiveEffects on items without
+    // the need to transfer the effect to the actor.
+    CONFIG.ActiveEffect.legacyTransferral = false;
+
     /**
     * Set an initiative formula for the system
     * @type {String}
@@ -282,38 +286,27 @@ Hooks.once("ready", async function () {
             ui.notifications.info(`Migragtion complete.`)
         }
 
-        // if lastMigration < 3.0.4-alpha
-        // if (foundry.utils.isNewerVersion('3.0.4', lastMigration)) {
-        //     ui.notifications.info(`Migragrating actor data.`)
-        //     for (let actor of game.actors.contents) {
-        //         for (let actor of game.actors) {
-        //             if (actor.statuses.has("dead")) {
-        //                 let statuses = actor.statuses
-        //                 statuses.delete("dead")
-        //                 statuses.add("defeated")
-        //                 await actor.update({ statuses: statuses })
-        //             }
-        //         }
-        //     }
-        //     ui.notifications.info(`Migragtion complete.`)
-        // }
-
+        // if lastMigration < 3.0.4
+        // Remove all tranferred effects
+        if (foundry.utils.isNewerVersion('3.0.4', lastMigration)) {
+            ui.notifications.info(`Migragrating actor data.`)
+            for (let actor of game.actors.contents) {
+                for( let effect of actor.effects.filter(o=> o.origin))
+                {
+                    effect.delete();
+                }
+            }
+            ui.notifications.info(`Migragtion complete.`)
+        }
+        
     }
-
-
-    //migrateWorld();
 
 });
 
-//const migrationTag = 20230611
 
 async function migrateRemoveDuplicateDefenseMovementItems() {
 
-    //if (game.actors.contents.find(o => o.system.migrationTag != migrationTag)) {
     ui.notifications.info(`Migragrating actor data.`)
-    // } else {
-    //     return;
-    // }
 
     let count = 0
     for (let actor of game.actors.contents) {
@@ -431,6 +424,9 @@ Hooks.on("renderDialog", (dialog, html, data) => {
 
 Hooks.on("renderActorSheet", (dialog, html, data) => {
     html.find('header h4').append(`<span>${game.system.version}<span>`)
+    // html.find('header h4').after(`<a class="header-button control configure-type">
+    // <i class="fal fa-user-robot"></i>Type 
+    // <a>`)
 })
 
 Hooks.on("renderItemSheet", (dialog, html, data) => {